Creates separates module for version 4 of apmecho.
